+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Flag to exclude content items from cached, S3 backed, CloudFront destinations using an object cache control header
 * 'Live' items for information updated in near real time (multiple times per hour)
 * Usage instructions for removing records as a last resort
+* Site checks for MAGIC Distribution Service hosted content
 * Azure Entra app registration for accessing MAGIC Distribution Service hosted content
 
 ### Fixed